refactor View/ViewEditor commands, fix animation bug on delete `undo()`
