+++ src/attack_flow_builder/src/assets/scripts/OpenChart/DiagramEditor/Commands/ViewEditor/Animations/SelectAnimation.ts
